Moved 'MedataManagerClient.py' to 'ESRF/ESRFMetadataManagerClient'. Moved metadata code from ID30A1MultiCollect to ESRFMultiCollect.
