+++ ESRF/ESRFMetadataManagerClient.py
